[Sofa.Component.Haptic] Rename into Sofa.Component.Haptics
