+++ Sofa/Component/Compat/Sofa.Component.Haptics.cmake
