+++ Sofa/Component/Compat/src/SofaHaptics/ForceFeedback.h
@@ -21,12 +21,12 @@
 ******************************************************************************/
 #pragma once
 
-#include <sofa/component/haptic/ForceFeedback.h>
+#include <sofa/component/haptics/ForceFeedback.h>
 
-// SOFA_DEPRECATED_HEADER("v22.06", "v23.06", "sofa/component/haptic/ForceFeedback.h")
+// SOFA_DEPRECATED_HEADER("v22.06", "v23.06", "sofa/component/haptics/ForceFeedback.h")
 
 namespace sofa::component::controller
 {
-    using ForceFeedback = sofa::component::haptic::ForceFeedback;
+    using ForceFeedback = sofa::component::haptics::ForceFeedback;
 
 } // namespace sofa::component::controller
