+++ Sofa/Component/Compat/src/SofaHaptics/NullForceFeedback.h
@@ -21,12 +21,12 @@
 ******************************************************************************/
 #pragma once
 
-#include <sofa/component/haptic/NullForceFeedback.h>
+#include <sofa/component/haptics/NullForceFeedback.h>
 
-// SOFA_DEPRECATED_HEADER("v22.06", "v23.06", "sofa/component/haptic/NullForceFeedback.h")
+// SOFA_DEPRECATED_HEADER("v22.06", "v23.06", "sofa/component/haptics/NullForceFeedback.h")
 
 namespace sofa::component::controller
 {
-    using NullForceFeedback = sofa::component::haptic::NullForceFeedback;
+    using NullForceFeedback = sofa::component::haptics::NullForceFeedback;
 
 } // namespace sofa::component::controller
