+++ Sofa/Component/Haptics/Sofa.Component.HapticsConfig.cmake.in
