+++ Sofa/Component/Haptics/src/sofa/component/haptics/NullForceFeedback.cpp
@@ -19,14 +19,29 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <sofa/component/haptic/NullForceFeedbackT.h>
+#include <sofa/component/haptics/NullForceFeedback.h>
+#include <sofa/core/visual/VisualParams.h>
 #include <sofa/core/ObjectFactory.h>
 
-namespace sofa::component::haptic
+namespace sofa::component::haptics
 {
 
-int nullForceFeedbackTClass = sofa::core::RegisterObject("Null force feedback for haptic feedback device")
-        .add< NullForceFeedbackT<sofa::defaulttype::Vec1Types> >()
-        .add< NullForceFeedbackT<sofa::defaulttype::Rigid3Types> >();
+void NullForceFeedback::init()
+{
+    this->ForceFeedback::init();
+}
 
-} // namespace sofa::component::haptic
+void NullForceFeedback::computeForce(SReal /*x*/, SReal /*y*/, SReal /*z*/, SReal /*u*/, SReal /*v*/, SReal /*w*/, SReal /*q*/, SReal& fx, SReal& fy, SReal& fz)
+{
+    fx = fy = fz = 0.0;
+}
+
+void NullForceFeedback::computeWrench(const sofa::defaulttype::SolidTypes<SReal>::Transform &/*world_H_tool*/, const sofa::defaulttype::SolidTypes<SReal>::SpatialVector &/*V_tool_world*/, sofa::defaulttype::SolidTypes<SReal>::SpatialVector &W_tool_world )
+{
+    W_tool_world.clear();
+}
+
+int nullForceFeedbackClass = sofa::core::RegisterObject("Null force feedback for haptic feedback device")
+        .add< NullForceFeedback >();
+
+} // namespace sofa::component::haptics
