+++ applications/plugins/SensableEmulation/CMakeLists.txt
@@ -15,7 +15,7 @@ set(SOURCE_FILES
 set(README_FILES PluginSensableEmulation.txt)
 
 find_package(Sofa.Config REQUIRED)
-sofa_find_package(Sofa.Component.Haptic REQUIRED)
+sofa_find_package(Sofa.Component.Haptics REQUIRED)
 sofa_find_package(Sofa.Component.Controller REQUIRED)
 sofa_find_package(Sofa.GL.Component.Rendering3D REQUIRED)
 sofa_find_package(Boost QUIET COMPONENTS thread REQUIRED)
@@ -24,7 +24,7 @@ include_directories(${Boost_INCLUDE_DIRS})
 
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES} ${README_FILES})
 set_target_properties(${PROJECT_NAME} PROPERTIES COMPILE_FLAGS "-DSOFA_BUILD_SENSABLEEMULATIONPLUGIN")
-target_link_libraries(${PROJECT_NAME} Sofa.Component.Haptic Sofa.Component.Controller Sofa.GL.Component.Rendering3D ${Boost_LIBRARIES})
+target_link_libraries(${PROJECT_NAME} Sofa.Component.Haptics Sofa.Component.Controller Sofa.GL.Component.Rendering3D ${Boost_LIBRARIES})
 
 include_directories("${CMAKE_CURRENT_SOURCE_DIR}/..")
