+++ modules/SofaHaptics/CMakeLists.txt
@@ -16,11 +16,11 @@ set(EXTRA_FILES
 
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES} ${EXTRA_FILES})
 
-message(WARNING "${PROJECT_NAME} module is deprecated. It will be removed at v23.06. Use Sofa.Component.Haptic instead.")
+message(WARNING "${PROJECT_NAME} module is deprecated. It will be removed at v23.06. Use Sofa.Component.Haptics instead.")
 
 # forward to the new sofang module
-sofa_find_package(Sofa.Component.Haptic REQUIRED)
-target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Haptic)
+sofa_find_package(Sofa.Component.Haptics REQUIRED)
+target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Component.Haptics)
 
 sofa_create_package_with_targets(
     PACKAGE_NAME ${PROJECT_NAME}
